requirements.txt
boto3==1.20.5
botocore==1.23.5
discord.py==1.7.3
moviepy==1.0.3
requests==2.26.0
requests-oauthlib==1.3.0
beautifulsoup4==4.10.0
python-dotenv==0.20.0
s3transfer==0.5.0